ENH: add mlflow support with a callback method for fitloop
